pyproject.toml
[tool.poetry]
name = "hw7"
version = "0.1.0"
description = ""
authors = ["A-Lastovets <164244571+A-Lastovets@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^26.1.0"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
